Add mesh gradient effect to Keyboard processing view text and icon
