+++ VivaDicta/Models/Transcription.swift
@@ -102,7 +102,7 @@ extension Transcription {
     [
         
         Transcription(
-            text: "hello world",
+            text: "hello world. heya how are you. asldjfsldafjasldfjsdf. Faldsjflasdjflsjadf. Flajsdflajsdflajsdfljasdflsaldfjlasdjflsadfjlasdjf. Flajlsdjfalsdjflajsdlfjasldfj. aFljalsdjflajsfdljasf",
             audioDuration: 5),
         
         Transcription(
